Fix warning: format '%u' expects type 'unsigned int', but argument 3 has type 'size_t'.
svn path=/trunk/netsurf/; revision=3264

+++ render/box.c
@@ -528,7 +528,7 @@ void box_dump(struct box *box, unsigned int depth)
 	}
 
 	if (box->text)
-		fprintf(stderr, "%u '%.*s' ", box->byte_offset,
+		fprintf(stderr, "%li '%.*s' ", (unsigned long) box->byte_offset,
 				(int) box->length, box->text);
 	if (box->space)
 		fprintf(stderr, "space ");
